--- MainTask/src/Bus.java
@@ -1,45 +1,0 @@
-
-public class Bus {
-    private String number;
-    private String model;
-    private int mileage;
-
-    Bus(String number, String model, int mileage) {
-        this.number = number;
-        this.model = model;
-        this.mileage = mileage;
-    }
-
-    public String getNumber() {
-        return number;
-    }
-
-    public void setNumber(String number) {
-        this.number = number;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-    public int getMileage() {
-        return mileage;
-    }
-
-    public void setMileage(int mileage) {
-        this.mileage = mileage;
-    }
-
-    @Override
-    public String toString() {
-        return "Bus{" +
-                "number='" + number + '\'' +
-                ", model='" + model + '\'' +
-                ", mileage=" + mileage +
-                '}';
-    }}
-
