+++ MainTask/src/models/Bus.java
@@ -1,0 +1,59 @@
+package models;
+
+public class Bus {
+    private final String number;
+    private final String model;
+    private final int mileage;
+
+    private Bus(String number, String model, int mileage) {
+        this.number = number;
+        this.model = model;
+        this.mileage = mileage;
+    }
+
+    public String getNumber() {
+        return number;
+    }
+
+    public String getModel() {
+        return model;
+    }
+
+    public int getMileage() {
+        return mileage;
+    }
+
+    @Override
+    public String toString() {
+        return "Bus{" +
+                "number='" + number + '\'' +
+                ", model='" + model + '\'' +
+                ", mileage=" + mileage +
+                '}';
+    }
+
+    public static class BusBuilder {
+        private String number;
+        private String model;
+        private int mileage;
+
+        public BusBuilder setNumber(String number) {
+            this.number = number;
+            return this;
+        }
+
+        public BusBuilder setModel(String model) {
+            this.model = model;
+            return this;
+        }
+
+        public BusBuilder setMileage(int mileage) {
+            this.mileage = mileage;
+            return this;
+        }
+
+        public Bus build() {
+            return new Bus(number, model, mileage);
+        }
+    }
+}
